udpating dtype on targets for one-hot encoding

+++ modules/eval_handler.py
@@ -43,7 +43,8 @@ class SequentialEvalLoop:
         sys.stdout.flush()
         
       # One-hot-ify targets
-      y = torch.eye(self.num_classes)[targets].type(torch.int64)
+      targets = targets.type(torch.int64)
+      y = torch.eye(self.num_classes)[targets]
       sample_count += y.shape[0]
 
       # Determine device placement
@@ -122,7 +123,8 @@ class CascadedEvalLoop(object):
         sys.stdout.write(f"\rBatch {batch_i+1}/{len(loader)}")
         sys.stdout.flush()
       # One-hot-ify targets
-      y = torch.eye(self.num_classes)[targets].type(torch.int64)
+      targets = targets.type(torch.int64)
+      y = torch.eye(self.num_classes)[targets]
       sample_count += y.shape[0]
 
       if self.keep_logits:
